Split contact related code into multiple files
